remote/client: move ensure_event_loop() to labgrid.util.loop
Moving the function and the ContextVar to labgrid.util.loop will allow
non-client code to use it as well. Otherwise usage will result in
circular imports such as:

      from labgrid import Target
  labgrid/__init__.py:1: in <module>
      from .target import Target
  labgrid/target.py:10: in <module>
      from .driver import Driver
  labgrid/driver/__init__.py:4: in <module>
      from .serialdriver import SerialDriver
  labgrid/driver/serialdriver.py:10: in <module>
      from ..util.proxy import proxymanager
  labgrid/util/proxy.py:7: in <module>
      from ..resource.common import Resource
  labgrid/resource/__init__.py:2: in <module>
      from .ethernetport import SNMPEthernetPort
  labgrid/resource/ethernetport.py:9: in <module>
      from ..remote.client import ensure_event_loop
  labgrid/remote/client.py:41: in <module>
      from .. import Environment, Target, target_factory
  E   ImportError: cannot import name 'Environment' from partially initialized module 'labgrid' (most likely due to a circular import) (labgrid/__init__.py)
